move target info to own file
